Push sp->resource FK down to control name
